add chain and model picking

+++ backend/models/models.py
@@ -13,7 +13,7 @@ class File(db.Model):
     name: str = db.Column(db.String(256), nullable=False)
     is_temp: bool = db.Column(db.Boolean, nullable=False, default=False)
     status: str = db.Column(db.String(16), nullable=False)
-    rmsd: int = db.Column(db.Integer)
+    rmsd: float = db.Column(db.Float)
     task_id: str = db.Column(
         db.String(TASK_ID_LENGTH), db.ForeignKey("task.id"), nullable=False
     )
